new lora frame, much smaller wow

+++ components/timers/mission_timer.h
@@ -6,7 +6,7 @@
 #include <stdbool.h>
 
 
-bool mission_timer_init(int64_t timer_disable_value);
+bool mission_timer_init(uint32_t timer_disable_value);
 
 /**
  * @brief Start mission timer
@@ -15,8 +15,8 @@ bool mission_timer_init(int64_t timer_disable_value);
  * @return true :D
  * @return false :C
  */
-bool mission_timer_start(int64_t countdown_begin_time_ms);
-int mission_timer_get_time();
+bool mission_timer_start(int32_t countdown_begin_time_ms);
+uint32_t mission_timer_get_time();
 bool mission_timer_is_enable();
 bool mission_timer_stop();
 bool mission_timer_set_disable_value(int64_t _disableValue);
