+++ components/utils/utils.c
@@ -2,5 +2,5 @@
 #include "esp_timer.h"
 
 uint32_t get_uptime_ms(void) {
-    return esp_timer_get_time() / 1000.0;
+    return (uint32_t)(esp_timer_get_time() / 1000);
 }
